BUG: Complete notebook-types refactoring by removing duplicate types.
